% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{addData.modelStrategy}
\alias{addData.modelStrategy}
\title{Add user data}
\usage{
\method{addData}{modelStrategy}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{xts}
}
\description{
Here user can add his own data in xts format
}
