% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{addData}
\alias{addData}
\title{Add user data}
\usage{
addData(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{named arguments}
}
\description{
Here user can add his own data in xts format
}
