% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R
\name{addDistribution.list}
\alias{addDistribution.list}
\title{Add distribution to list of models}
\usage{
\method{addDistribution}{list}(l, ...)
}
\arguments{
\item{l}{list, list of modelStrategy objects}

\item{...}{params for addDistribution}
}
\description{
This method add the same distribution to each model in list
}
