% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{addDistribution.modelPortfolio}
\alias{addDistribution.modelPortfolio}
\title{Add distribution to list of models}
\usage{
\method{addDistribution}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params for addDistribution}
}
\description{
This method add the same distribution to each model in list
}
