% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R, R/modelPortfolio.R
\name{addDistribution}
\alias{addDistribution}
\alias{addDistribution.modelStrategy}
\alias{addDistribution.list}
\alias{addDistribution.modelPortfolio}
\title{Add distribution}
\usage{
addDistribution(this, paramset.label, component.type, component.label,
  variable, label)

\method{addDistribution}{modelStrategy}(this, paramset.label,
  component.type, component.label, variable, label)

\method{addDistribution}{list}(this, ...)

\method{addDistribution}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{model}

\item{paramset.label}{character, label of paramset, this argument can be missed}

\item{component.type}{character, one of the followng names: rule, indicator, params, lookback, lookforward, beta_fun}

\item{component.label}{character, name of component, argument 'as' is resposible for that. 
If component.type is equal to one of lookback, lookforward, beta_fun, then this argument should be missed}

\item{variable}{list with only one element, for example list(n = 1:10)}

\item{label}{character, name for this distribution}

\item{...}{params for addDistribution}
}
\description{
This method add the same distribution to each model in list
}
\examples{
\dontrun{
addIndicator(this, args = list(name = BBands, HLC = quote(spread), n = 20, sd = 1), as = 'bb')
addDistribution(this,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread,1) > Lag(bb[ ,'up'],1)) &
                  (spread < bb[, 'up']) &
                  (spread > bb[ ,'mavg']) &
                  (abs(spread - bb[, 'mavg'])  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short')
addDistribution(this,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)
}
}
