% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{addDistributionConstraint.modelPortfolio}
\alias{addDistributionConstraint.modelPortfolio}
\title{Add distributions' constraint to list of models}
\usage{
\method{addDistributionConstraint}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params for addDistribution}
}
\description{
This method add the same distributions' constraint to each model in list
}
