% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{addDocker}
\alias{addDocker}
\title{Add dockerfile and create docker container on the server}
\usage{
addDocker(path, dockername, session, verbose = FALSE)
}
\arguments{
\item{path}{character, path to dockerfile}

\item{dockername}{character, name of docker container}

\item{session}{ssh session}

\item{verbose}{logical}
}
\value{
character, an answer from the server
}
\description{
Your dockerfile should be with installed python and pip
}
