% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{addIndicator}
\alias{addIndicator}
\alias{addIndicator.modelStrategy}
\title{Adds indicators to strategy}
\usage{
addIndicator(this, args, as, lookback, ...)

\method{addIndicator}{modelStrategy}(this, args, as, lookback, ...)
}
\arguments{
\item{this}{model}

\item{args}{list, the first argument should be name, it should be equal to function (your indicator), other arguments of list
will be passed to this function}

\item{as}{character, local name of this indicator}

\item{lookback}{numeric, how many points does it need to do calculation}

\item{...}{addtional params}
}
\description{
This indicator will be calculated each time when spread update coefs
indicators must return data.frame or matrix or array
}
\examples{
\dontrun{
addIndicator(this, args = list(name = BBands, HLC = quote(spread), n = 20, sd = 1), as = 'bb',
    lookback = 100)
}
}
