% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericModelStrategy.R, R/rules_indicators.R
\name{addIndicators}
\alias{addIndicators}
\alias{addIndicators.modelStrategy}
\title{Add multiple indicators to strategy}
\usage{
addIndicators(this, ...)

\method{addIndicators}{modelStrategy}(this, expr, names, as,
  lookback = 0, args = list(), ...)
}
\arguments{
\item{this}{model}

\item{...}{params}

\item{expr}{expression}

\item{names}{character vector, names of indicators to include in.}

\item{as}{character, name of set of indicators}

\item{lookback}{numeric, periods of time for calculation}

\item{args}{list, arguments of expression}
}
\description{
These indicators will be calculated each time when spread updates coefficients
indicators must return data.frame or matrix or array
}
