% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{addObject}
\alias{addObject}
\alias{addObject.modelStrategy}
\title{Add user-defined objects to modelStrategy for future usage in backtest}
\usage{
addObject(this, ...)

\method{addObject}{modelStrategy}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{named args}
}
\description{
Add user-defined objects to modelStrategy for future usage in backtest
}
