% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionManagers.R
\name{addPM.modelStrategy}
\alias{addPM.modelStrategy}
\title{Creater for position manager.}
\usage{
\method{addPM}{modelStrategy}(this, as, increase = list(cond =
  quote(FALSE), expr = quote({ })), decrease = list(cond = quote(FALSE),
  expr = quote({ })), rebalance = list(cond = quote(FALSE), expr =
  quote({ })), change = list(cond = quote(FALSE), expr = quote({ })),
  close = list(cond = quote(FALSE)), oco = "base", args = list())
}
\arguments{
\item{this}{modelStrategy}

\item{as}{character, name}

\item{increase}{list, the first argument is resposible for condition when this rule will work,
the second argument is expression that must calculate size of position(mult_delta)(how many spreads it needs to be bought).
In another words the last row must contain 'mult_delta <- ...'}

\item{decrease}{list, see increase, in expr quote beta_new must be defined}

\item{rebalance}{list, see increase, in expr quote beta_new must be defined}

\item{change}{list, see increase, in expr quote mult_delta must be defined}

\item{close}{list, see increase}

\item{oco}{charcter, name of namespace}

\item{args}{list, arguments of that manager}
}
\description{
Creater for position manager.
}
