% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{addProgramPart.modelStrategy}
\alias{addProgramPart.modelStrategy}
\title{Adds variables to strategy}
\usage{
\method{addProgramPart}{modelStrategy}(this, as, evolution = list())
}
\arguments{
\item{this}{modelStrategy}

\item{as}{character, name}

\item{evolution}{list, name of each element is the place where code will be executed, and value is quoted expression
There are several places :
1. init -- in initialization period before main cycle
2. after_tables -- after initialization of rules tables
3. after_coefs -- after initialization of coeffitients
4. each_iter -- in the beginnig of each iteration, before execution of rules
5. after_enter_to_pos -- Right after entering in position
6. unrealized_money_last -- here you can change evolution of unrealized pnl. in this quote must be following piece of code
unrealized_money_last <-  <YOUR CODE HERE>
7. before_enter_to_pos -- it executes just before entering to position
8. befor_exit_from_pos -- it executes just before exiting from position
9. data -- in this section you can define your datasets, that depend on modelData slots}
}
\description{
Adds variables to strategy
}
