% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{addStat.modelStrategy}
\alias{addStat.modelStrategy}
\title{This variable will be included to program and after calling perform function  it will be updated}
\usage{
\method{addStat}{modelStrategy}(this, name, value)
}
\arguments{
\item{this}{modelStrategy}

\item{name}{character}

\item{value}{not NULL}
}
\description{
This variable will be included to program and after calling perform function  it will be updated
}
