% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{addTradeTime.modelStrategy}
\alias{addTradeTime.modelStrategy}
\title{Adds time when trading is permitted}
\usage{
\method{addTradeTime}{modelStrategy}(this, type, l)
}
\arguments{
\item{this}{modelStrategy}

\item{type}{character, one of enter, open, exit, close}

\item{l}{list, each element is a string that represents time}
}
\description{
Adds time when trading is permitted
}
