% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{applyParamsetServer.list}
\alias{applyParamsetServer.list}
\title{Apply paramset to list of strategies}
\usage{
\method{applyParamsetServer}{list}(l, session, paramset.label, nsamples = 100,
  verbose = FALSE, ...)
}
\arguments{
\item{l, }{list of modelStrategy}

\item{session}{ssh_session}

\item{paramset.label}{charater, name of paramset}

\item{nsamples}{numeric, how many backtests to do}

\item{start_date}{Date / charater in format 'yyyy-mm-dd', start date of simulation}

\item{end_date}{Date / charater in format 'yyyy-mm-dd', end date of simulation}
}
\value{
list, list of data.frames (reports of simulations)
}
\description{
Apply paramset to list of strategies
}
