% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{applyParamsetServer.modelPortfolio}
\alias{applyParamsetServer.modelPortfolio}
\title{Apply paramset to list of strategies and summarize results}
\usage{
\method{applyParamsetServer}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{arguments to perform function}
}
\description{
Apply paramset to list of strategies and summarize results
}
