% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/user_communication.R
\name{applyParamsetServer.modelPortfolio}
\alias{applyParamsetServer.modelPortfolio}
\alias{applyParamsetServer}
\alias{applyParamsetServer.modelStrategy}
\alias{applyParamsetServer.list}
\title{Do backtest of combinations from paramset}
\usage{
\method{applyParamsetServer}{modelPortfolio}(this, ...)

applyParamsetServer(this, paramset.label, session, nsamples = 100,
  verbose = FALSE, ...)

\method{applyParamsetServer}{modelStrategy}(this, paramset.label, session,
  nsamples = 100, verbose = FALSE, ...)

\method{applyParamsetServer}{list}(this, paramset.label, session,
  nsamples = 100, verbose = FALSE, ...)
}
\arguments{
\item{this}{model}

\item{...}{additional arguments (start_date, end_date)}

\item{paramset.label}{charater, name of paramset}

\item{session}{ssh_session}

\item{nsamples}{numeric, how many backtests to do}

\item{verbose}{logical, if true then logs will be printed}
}
\value{
data.frame, reports of simulations

list, list of data.frames (reports of simulations)
}
\description{
Do backtest of combinations from paramset
}
