% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{clearData}
\alias{clearData}
\alias{clearData.modelStrategy}
\title{Clear data that has been specified before}
\usage{
clearData(this)

\method{clearData}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Clear data that has been specified before
}
