% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{cloneStrategy.modelStrategy}
\alias{cloneStrategy.modelStrategy}
\alias{cloneStrategy.modelPortfolio}
\alias{cloneStrategy}
\title{Copy modelStrategy object}
\usage{
\method{cloneStrategy}{modelStrategy}(this, clone_data = FALSE,
  clone_backtests = FALSE, ...)

\method{cloneStrategy}{modelPortfolio}(this, ...)

cloneStrategy(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{clone_data}{logical, if TRUE then clone modelData object}

\item{clone_backtests}{logical, if TRUE then all backtests will be cloned}

\item{...}{params}
}
\value{
modelStrategy

modelPortfolio
}
\description{
Copy modelStrategy object
}
