% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{deleteOldData}
\alias{deleteOldData}
\alias{deleteOldData.modelStrategy}
\alias{deleteOldData.list}
\alias{deleteOldData.modelPortfolio}
\title{Delete backtest data from model}
\usage{
deleteOldData(this)

\method{deleteOldData}{modelStrategy}(this)

\method{deleteOldData}{list}(this)

\method{deleteOldData}{modelPortfolio}(this)
}
\arguments{
\item{this}{model}
}
\description{
Delete backtest data from model
}
