% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R
\name{deleteParamset.list}
\alias{deleteParamset.list}
\title{Removes paramset from each strategy in list of models}
\usage{
\method{deleteParamset}{list}(l, ...)
}
\arguments{
\item{l}{list, list of modelStrategy objects}

\item{...}{params for addDistribution}
}
\description{
Removes paramset from each strategy in list of models
}
