% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{deleteParamset.modelPortfolio}
\alias{deleteParamset.modelPortfolio}
\title{Removes paramset from each strategy in list of models}
\usage{
\method{deleteParamset}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params for addDistribution}
}
\description{
Removes paramset from each strategy in list of models
}
