% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R, R/modelPortfolio.R
\name{deleteParamset}
\alias{deleteParamset}
\alias{deleteParamset.modelStrategy}
\alias{deleteParamset.list}
\alias{deleteParamset.modelPortfolio}
\title{Remove paramset from strategy}
\usage{
deleteParamset(this, paramset.label)

\method{deleteParamset}{modelStrategy}(this, paramset.label)

\method{deleteParamset}{list}(this, ...)

\method{deleteParamset}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{model}

\item{paramset.label}{name of paramset}

\item{...}{params for addDistribution}
}
\description{
Remove paramset from strategy

Remove paramset from each strategy in list of models
}
