% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymodel.R
\name{deletePyModel.modelStrategy}
\alias{deletePyModel.modelStrategy}
\title{Delete settings of python model}
\usage{
\method{deletePyModel}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Delete settings of python model
}
