% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherHelpers.R
\name{Diff}
\alias{Diff}
\alias{Diff.default}
\alias{Diff.matrix}
\title{Work as function diff, but remains NA in the beginnig.}
\usage{
Diff(x, ...)

\method{Diff}{default}(x, k = 1)

\method{Diff}{matrix}(x, k = 1)
}
\arguments{
\item{x}{object}

\item{...}{params}

\item{k}{numeric, integer lag}
}
\description{
Work as function diff, but remains NA in the beginnig.
}
