% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_examples.R
\name{download_examples}
\alias{download_examples}
\title{Downloads directory of examples of some user}
\usage{
download_examples(session, dir = "main", to = getwd())
}
\arguments{
\item{session}{ssh_session}

\item{dir}{character, name of dir, by default it equals to main}

\item{to}{character, path where you want to store examples}
}
\description{
Downloads directory of examples of some user
}
