% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.modelStrategy}
\title{Evaluate strategy in contest}
\usage{
evaluate(this, contest, session, verbose = FALSE)

\method{evaluate}{modelStrategy}(this, contest, session, verbose = FALSE)
}
\arguments{
\item{this}{modelStrategy}

\item{contest}{character, name of contest}

\item{session}{ssh_session}

\item{verbose}{logical}
}
\description{
Evaluate strategy in contest
}
