% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{extractStat.modelStrategy}
\alias{extractStat.modelStrategy}
\title{Updates environment}
\usage{
\method{extractStat}{modelStrategy}(this, e)
}
\arguments{
\item{this}{modelStrategy}

\item{e}{environment}
}
\description{
Updates environment
}
