% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getBacktestResults.modelPortfolio}
\alias{getBacktestResults.modelPortfolio}
\title{Results of backtest}
\usage{
\method{getBacktestResults}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, name of paramset}
}
\value{
list of data.frames
}
\description{
Results of backtest
}
