% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getBacktestResults.modelStrategy}
\alias{getBacktestResults.modelStrategy}
\title{Results of backtest}
\usage{
\method{getBacktestResults}{modelStrategy}(this, paramset.label)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, name of paramset}
}
\value{
data.frame
}
\description{
Results of backtest
}
