% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getBacktestResults}
\alias{getBacktestResults}
\alias{getBacktestResults.modelStrategy}
\alias{getBacktestResults.list}
\alias{getBacktestResults.modelPortfolio}
\title{Results of backtest}
\usage{
getBacktestResults(this, paramset.label)

\method{getBacktestResults}{modelStrategy}(this, paramset.label)

\method{getBacktestResults}{list}(this, ...)

\method{getBacktestResults}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, name of paramset}

\item{...}{arguments to getBacktestResults.modelStrategy}
}
\value{
data.frame

list of data.frames

list of data.frames
}
\description{
Results of backtest
}
