% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getBetaData}
\alias{getBetaData}
\alias{getBetaData.modelStrategy}
\alias{setBetaData}
\alias{setBetaData.modelStrategy}
\title{Get/set name of table for calculating coefficients.}
\usage{
getBetaData(this)

\method{getBetaData}{modelStrategy}(this)

setBetaData(this, x)

\method{setBetaData}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{character, name of table in modelD object}
}
\description{
Get/set name of table for calculating coefficients.
}
