% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getBetasInt.modelStrategy}
\alias{getBetasInt.modelStrategy}
\title{Return if you have a need to convert betas to int.
It is TRUE by default}
\usage{
\method{getBetasInt}{modelStrategy}(model)
}
\arguments{
\item{model}{modelStrategy}
}
\description{
Return if you have a need to convert betas to int.
It is TRUE by default
}
