% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getBetasInt}
\alias{getBetasInt}
\alias{getBetasInt.modelStrategy}
\title{Return if you have a need to convert betas to int.
It is TRUE by default}
\usage{
getBetasInt(this)

\method{getBetasInt}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Return if you have a need to convert betas to int.
It is TRUE by default
}
