% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCapital}
\alias{getCapital}
\alias{getCapital.modelStrategy}
\alias{getCapital.modelPortfolio}
\title{Return money in position}
\usage{
getCapital(this, ...)

\method{getCapital}{modelStrategy}(this, ...)

\method{getCapital}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}
}
\description{
Return money in position
}
