% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDateByIndex}
\alias{getDateByIndex}
\alias{getDateByIndex.modelStrategy}
\alias{getDateByIndex.modelPortfolio}
\title{Get dates by indexes from modelData}
\usage{
getDateByIndex(this, indexes = NULL)

\method{getDateByIndex}{modelStrategy}(this, indexes = NULL)

\method{getDateByIndex}{modelPortfolio}(this, indexes = NULL)
}
\arguments{
\item{this}{modelStrategy}

\item{indexes}{numeric vector or NULL}
}
\value{
vector of dates
}
\description{
Get dates by indexes from modelData
}
