% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_contraints.R
\name{getDistributions.modelStrategy}
\alias{getDistributions.modelStrategy}
\title{Gets list of distributions}
\usage{
\method{getDistributions}{modelStrategy}(this, paramset.label)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character}
}
\value{
list of distributions
}
\description{
Gets list of distributions
}
