% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R
\name{getDistributions}
\alias{getDistributions}
\alias{getDistributions.modelStrategy}
\title{Get list of distributions}
\usage{
getDistributions(this, paramset.label)

\method{getDistributions}{modelStrategy}(this, paramset.label)
}
\arguments{
\item{this}{model}

\item{paramset.label}{character}
}
\value{
list of distributions
}
\description{
Get list of distributions
}
