% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericModelStrategy.R, R/modelStrategy.R
\name{getExpandingLookback}
\alias{getExpandingLookback}
\alias{getExpandingLookback.modelStrategy}
\title{get expandingLookback}
\usage{
getExpandingLookback(this)

\method{getExpandingLookback}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
If it is TRUE, then size of data in beta_fun will be equal to current index, else to lookback
}
