% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{getIndicators.modelStrategy}
\alias{getIndicators.modelStrategy}
\title{Gets inforamtion about indicators in the model}
\usage{
\method{getIndicators}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list of inforamation
}
\description{
Gets inforamtion about indicators in the model
}
