% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getLookback.modelStrategy}
\alias{getLookback.modelStrategy}
\title{Gets window for calculating coefs}
\usage{
\method{getLookback}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets window for calculating coefs
}
