% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getMaxLookback}
\alias{getMaxLookback}
\alias{getMaxLookback.modelStrategy}
\title{Get window for calculating indicators}
\usage{
getMaxLookback(this)

\method{getMaxLookback}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Get window for calculating indicators
}
