% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/utils.R
\name{getModelD.modelPortfolio}
\alias{getModelD.modelPortfolio}
\alias{getModelD.modelPortfolio}
\title{Get modelData}
\usage{
\method{getModelD}{modelPortfolio}(this)

\method{getModelD}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelPortfolio}

\item{this}{modelPortfolio}
}
\description{
Get modelData

Get modelData
}
