% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/utils.R
\name{getModelD.modelPortfolio}
\alias{getModelD.modelPortfolio}
\alias{getModelD.list}
\alias{getModelD}
\alias{getModelD.modelStrategy}
\title{Get modelData object from modelStrategy object}
\usage{
\method{getModelD}{modelPortfolio}(this)

\method{getModelD}{list}(this, ...)

getModelD(this)

\method{getModelD}{modelStrategy}(this)

\method{getModelD}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{arguments}
}
\value{
list of modelDatas
}
\description{
Get modelData object from modelStrategy object
}
