% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{getMoney.modelPortfolio}
\alias{getMoney.modelPortfolio}
\title{Return sum money of models inside portfolio}
\usage{
\method{getMoney}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelPortfolio}
}
\value{
numeric
}
\description{
Return sum money of models inside portfolio
}
