% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/modelStrategy.R
\name{getMoney.modelPortfolio}
\alias{getMoney.modelPortfolio}
\alias{getMoney}
\title{Return sum money of models inside portfolio}
\usage{
\method{getMoney}{modelPortfolio}(this)

getMoney(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
numeric, amount of money
}
\description{
Return sum money of models inside portfolio

Get amount of money that we have at the beginning in specific backtest
}
