% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionManagers.R
\name{getPM}
\alias{getPM}
\alias{getPM.modelStrategy}
\title{Gets all position managers}
\usage{
getPM(this, recalc, ...)

\method{getPM}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}

\item{recalc}{logical}

\item{...}{params}
}
\value{
list
}
\description{
Gets all position managers
}
