% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getPmAfterOpen}
\alias{getPmAfterOpen}
\alias{getPmAfterOpen.modelStrategy}
\title{Get pmAfterOpen.}
\usage{
getPmAfterOpen(this)

\method{getPmAfterOpen}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
If it is true, then position manager will be turned on just after opening position
}
