% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPnL.list}
\alias{getPnL.list}
\title{Get pnl as list of xts series}
\usage{
\method{getPnL}{list}(l)
}
\arguments{
\item{this}{modelPortfolio}
}
\value{
xts zoo
}
\description{
This method should be called only after backtest is done
}
