% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPnL.modelStrategy}
\alias{getPnL.modelStrategy}
\title{Get pnl as xts series}
\usage{
\method{getPnL}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
xts zoo
}
\description{
This method should be called only after backtest is done
}
