% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPnL}
\alias{getPnL}
\alias{getPnL.modelStrategy}
\alias{getPnL.modelPortfolio}
\alias{getPnL.list}
\title{Get pnl as xts series}
\usage{
getPnL(this)

\method{getPnL}{modelStrategy}(this)

\method{getPnL}{modelPortfolio}(this)

\method{getPnL}{list}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
xts zoo

xts zoo
}
\description{
This method should be called only after backtest is done
}
