% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{getProgramParts.modelStrategy}
\alias{getProgramParts.modelStrategy}
\title{Gets list of variables(program parts)}
\usage{
\method{getProgramParts}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets list of variables(program parts)
}
