% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{getProgramParts}
\alias{getProgramParts}
\alias{getProgramParts.modelStrategy}
\title{Get list of variables(program parts)}
\usage{
getProgramParts(this)

\method{getProgramParts}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Get list of variables(program parts)
}
