% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{getPublicName}
\alias{getPublicName}
\alias{getPublicName.modelStrategy}
\title{Get public name of user. It will be defined after you submit this model}
\usage{
getPublicName(this)

\method{getPublicName}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Get public name of user. It will be defined after you submit this model
}
