% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymodel.R
\name{getPyModel}
\alias{getPyModel}
\alias{getPyModel.modelStrategy}
\title{Get settings of python model}
\usage{
getPyModel(this)

\method{getPyModel}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list
}
\description{
Get settings of python model
}
