% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportCalendar.list}
\alias{getReportCalendar.list}
\title{Report of strategy year by year}
\usage{
\method{getReportCalendar}{list}(l)
}
\arguments{
\item{l}{list of modelStrategies}
}
\description{
Report of strategy year by year
}
