% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportCalendar.modelPortfolio}
\alias{getReportCalendar.modelPortfolio}
\title{Report of strategy year by year}
\usage{
\method{getReportCalendar}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelPortfolio}
}
\description{
Report of strategy year by year
}
