% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportCalendar.modelStrategy}
\alias{getReportCalendar.modelStrategy}
\title{Report of strategy year by year}
\usage{
\method{getReportCalendar}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Report of strategy year by year
}
