% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportCalendar}
\alias{getReportCalendar}
\alias{getReportCalendar.modelStrategy}
\alias{getReportCalendar.list}
\alias{getReportCalendar.modelPortfolio}
\title{Report of strategy year by year}
\usage{
getReportCalendar(this)

\method{getReportCalendar}{modelStrategy}(this)

\method{getReportCalendar}{list}(this)

\method{getReportCalendar}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list of data.frames
}
\description{
Report of strategy year by year
}
