% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportStrategy.list}
\alias{getReportStrategy.list}
\title{Report of strategy}
\usage{
\method{getReportStrategy}{list}(l)
}
\arguments{
\item{l}{list of modelStrategies}
}
\description{
Report of strategy
}
