% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportStrategy.modelPortfolio}
\alias{getReportStrategy.modelPortfolio}
\title{Report of strategy}
\usage{
\method{getReportStrategy}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelPortfolio}
}
\description{
Report of strategy
}
