% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportStrategy.modelStrategy}
\alias{getReportStrategy.modelStrategy}
\title{Report of strategy}
\usage{
\method{getReportStrategy}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Report of strategy
}
