% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportStrategy}
\alias{getReportStrategy}
\alias{getReportStrategy.modelStrategy}
\alias{getReportStrategy.list}
\alias{getReportStrategy.modelPortfolio}
\title{Report of strategy}
\usage{
getReportStrategy(this)

\method{getReportStrategy}{modelStrategy}(this)

\method{getReportStrategy}{list}(this)

\method{getReportStrategy}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
noquote data.frame with report

data.frame
}
\description{
Report of strategy
}
