% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportTrades.modelPortfolio}
\alias{getReportTrades.modelPortfolio}
\title{Report of trades}
\usage{
\method{getReportTrades}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelPortfolio}
}
\description{
Report of trades
}
