% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportTrades.modelStrategy}
\alias{getReportTrades.modelStrategy}
\title{Report of trades}
\usage{
\method{getReportTrades}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Report of trades
}
