% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportTrades}
\alias{getReportTrades}
\alias{getReportTrades.modelStrategy}
\alias{getReportTrades.list}
\alias{getReportTrades.modelPortfolio}
\title{Report of trades}
\usage{
getReportTrades(this)

\method{getReportTrades}{modelStrategy}(this)

\method{getReportTrades}{list}(this)

\method{getReportTrades}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Report of trades
}
