% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{getRules}
\alias{getRules}
\alias{getRules.modelStrategy}
\title{Get inforamtion about rules in the model}
\usage{
getRules(this)

\method{getRules}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list of inforamation
}
\description{
Get inforamtion about rules in the model
}
