% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{getScores}
\alias{getScores}
\alias{getScores.modelStrategy}
\title{Get scores from submitted model}
\usage{
getScores(this)

\method{getScores}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list
}
\description{
Get scores from submitted model
}
