% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getSpreadData}
\alias{getSpreadData}
\alias{getSpreadData.modelStrategy}
\alias{setSpreadData}
\alias{setSpreadData.modelStrategy}
\title{Get/set name of table for building spread.}
\usage{
getSpreadData(this)

\method{getSpreadData}{modelStrategy}(this)

setSpreadData(this, x)

\method{setSpreadData}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{character, name of table in modelD object}
}
\description{
Get/set name of table for building spread.
}
