% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getToleranceBeta.modelStrategy}
\alias{getToleranceBeta.modelStrategy}
\title{Gets tolerance of betas computation}
\usage{
\method{getToleranceBeta}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
numeric
}
\description{
Gets tolerance of betas computation
}
