% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getTradingData.modelStrategy}
\alias{getTradingData.modelStrategy}
\title{tradingData indicates which data strategy should use for spread calculating}
\usage{
\method{getTradingData}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
character
}
\description{
tradingData indicates which data strategy should use for spread calculating
}
