% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getWaitAfterClose.modelStrategy}
\alias{getWaitAfterClose.modelStrategy}
\title{Gets waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time}
\usage{
\method{getWaitAfterClose}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time
}
