% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getWaitAfterClose}
\alias{getWaitAfterClose}
\alias{getWaitAfterClose.modelStrategy}
\title{Get waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time}
\usage{
getWaitAfterClose(this)

\method{getWaitAfterClose}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Get waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time
}
