% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{interruptSimulation}
\alias{interruptSimulation}
\title{Interrupt simulation on the server}
\usage{
interruptSimulation(session, verbose = TRUE)
}
\arguments{
\item{session}{ssh_session}

\item{verbose}{logical, print message or not}
}
\description{
Interrupt simulation on the server
}
