% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{modelPortfolio}
\alias{modelPortfolio}
\title{creates modelPortfolio object}
\usage{
modelPortfolio(...)
}
\arguments{
\item{...}{list or models, list of models}
}
\value{
modelPortfolio object
}
\description{
creates modelPortfolio object
}
