% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_examples.R
\name{peek_examples}
\alias{peek_examples}
\title{Peek examples}
\usage{
peek_examples(session)
}
\arguments{
\item{session}{ssh_session}
}
\value{
list
}
\description{
Peek how many examples each user have
}
