% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{performServer.modelPortfolio}
\alias{performServer.modelPortfolio}
\title{Simulate portfolio of strategies}
\usage{
\method{performServer}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{arguments to perform function}
}
\description{
Simulate portfolio of strategies
}
