% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/user_communication.R
\name{performServer.modelPortfolio}
\alias{performServer.modelPortfolio}
\alias{performServer}
\alias{performServer.modelStrategy}
\alias{performServer.list}
\title{Send your strategy to server for backtesting.}
\usage{
\method{performServer}{modelPortfolio}(this, ...)

performServer(this, session, verbose = FALSE, ...)

\method{performServer}{modelStrategy}(this, session, verbose = FALSE,
  ...)

\method{performServer}{list}(this, session, verbose = FALSE, ...)
}
\arguments{
\item{this}{model}

\item{...}{additional arguments:
start_date -- from this date backtest will be started
end_date -- at this date backtest will be ended. Max end_date is Sys.Date() - 365
paramset.label -- character, name of paramset
paramset.index -- numeric, index of combination of params to backtest, by defualt current strategy will be backtested}

\item{session}{ssh_session}

\item{verbose}{logical, if true then logs will be printed}
}
\value{
data.frame, report of strategy

list
}
\description{
All results could be seen in your folder on server.
}
