% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCalendar.modelPortfolio}
\alias{plotCalendar.modelPortfolio}
\title{plot pnl in month-year matrix}
\usage{
\method{plotCalendar}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params}
}
\description{
plot pnl in month-year matrix
}
