% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCalendar.modelStrategy}
\alias{plotCalendar.modelStrategy}
\title{Plot pnl in month-year matrix}
\usage{
\method{plotCalendar}{modelStrategy}(this, compounded = FALSE)
}
\arguments{
\item{this}{modelStrategy}

\item{compounded}{logical, compounded returns to use or not}

\item{from}{character, name of backtest}
}
\description{
Plot pnl in month-year matrix
}
