% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCalendar}
\alias{plotCalendar}
\alias{plotCalendar.modelStrategy}
\alias{plotCalendar.xts}
\alias{plotCalendar.modelPortfolio}
\title{Plot pnl in month-year matrix}
\usage{
plotCalendar(this, ...)

\method{plotCalendar}{modelStrategy}(this, compounded = FALSE, ...)

\method{plotCalendar}{xts}(this, ...)

\method{plotCalendar}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{compounded}{logical, compounded returns to use or not}
}
\description{
Plot pnl in month-year matrix
}
