% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCalendar.xts}
\alias{plotCalendar.xts}
\title{Plot pnl in month-year matrix}
\usage{
\method{plotCalendar}{xts}(x)
}
\arguments{
\item{x}{xts one column of cumulative profit and loss}
}
\description{
Plot pnl in month-year matrix
}
