% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCapital}
\alias{plotCapital}
\alias{plotCapital.modelStrategy}
\alias{plotCapital.modelPortfolio}
\title{Plot Capital of strategy}
\usage{
plotCapital(this, ...)

\method{plotCapital}{modelStrategy}(this, interactive_plot = TRUE, ...)

\method{plotCapital}{modelPortfolio}(this, interactive_plot = TRUE, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{interactive_plot, }{logical, use FALSE if you want to disable the interactive graph}
}
\description{
Plot Capital of strategy
}
