% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDrawdowns.list}
\alias{plotDrawdowns.list}
\title{plot profit and loss graph}
\usage{
\method{plotDrawdowns}{list}(l, legend = TRUE, ...)
}
\arguments{
\item{l}{list, list of strategies}

\item{...}{params}
}
\description{
plot profit and loss graph
}
