% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDrawdowns.modelPortfolio}
\alias{plotDrawdowns.modelPortfolio}
\title{plot drawdowns}
\usage{
\method{plotDrawdowns}{modelPortfolio}(this, legend = TRUE, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params}
}
\description{
plot drawdowns
}
