% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDrawdowns.modelStrategy}
\alias{plotDrawdowns.modelStrategy}
\title{Plot drawdowns}
\usage{
\method{plotDrawdowns}{modelStrategy}(this, from = "base",
  return_type = "plot", graph_type = "ggplot2")
}
\arguments{
\item{this}{modelStrategy}

\item{from}{character, name of backtest}

\item{return_type}{character, plot or data}

\item{graph_type}{character, ggplot2 or xts}
}
\value{
ggplot/xts
}
\description{
Plot drawdowns
}
