% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPnL.modelPortfolio}
\alias{plotPnL.modelPortfolio}
\title{plot profit and loss graph}
\usage{
\method{plotPnL}{modelPortfolio}(this, legend = TRUE, ...)
}
\arguments{
\item{this}{modelPortfolio}

\item{...}{params}
}
\description{
plot profit and loss graph
}
