% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotReturns}
\alias{plotReturns}
\title{Plot graph maximum/minimum pnl in trade vs realized pnl}
\usage{
plotReturns(report, type = "MAE")
}
\arguments{
\item{report}{data.frame, must contain 3 columns 'MAE.with.com', 'MFE.with.com', 'pnl.sum.adj'. 
It can be report of trades, returned from simulation}

\item{type}{character, 'MFE' or 'MAE'}
}
\value{
ggplot
}
\description{
Plot graph maximum/minimum pnl in trade vs realized pnl
}
