% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotShiny}
\alias{plotShiny}
\alias{plotShiny.modelStrategy}
\title{Plot interactive distribution params}
\usage{
plotShiny(this, ...)

\method{plotShiny}{modelStrategy}(this, session, paramset = 1, ...)
}
\arguments{
\item{this}{it is Strategy}

\item{...}{params for shinyApp}

\item{session}{object of class ssh_session}

\item{paramset}{name of paramset}
}
\description{
Plot interactive distribution params
}
