% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{printTradeTime}
\alias{printTradeTime}
\alias{printTradeTime.modelStrategy}
\title{Prints trading time in format of time}
\usage{
printTradeTime(this, type = "all")

\method{printTradeTime}{modelStrategy}(this, type = "all")
}
\arguments{
\item{this}{model}

\item{type}{character, one of c('all', 'enter', 'open', 'exit', 'close')}
}
\description{
Prints trading time in format of time
}
