% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOsFun.R
\name{proportionOs}
\alias{proportionOs}
\title{Function for volume in rule}
\usage{
proportionOs(beta, money, proportion, price, ...)
}
\arguments{
\item{beta}{numeric vector, coefficients of spread}

\item{money}{numeric, all money that strategy has}

\item{proportion}{numeric, number between 0 and 1, how much money we can use in rule}

\item{price}{current price of instruments}

\item{...}{additional args}
}
\value{
numeric multiplier( how much minimal position to buy)
}
\description{
How many sum(price \* abs(beta)) in money \* proportion
}
