% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{reinitStat.modelStrategy}
\alias{reinitStat.modelStrategy}
\title{Reinit all variables in stats}
\usage{
\method{reinitStat}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Reinit all variables in stats
}
