% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{removeIndicator.modelStrategy}
\alias{removeIndicator.modelStrategy}
\title{Removes indicator by name}
\usage{
\method{removeIndicator}{modelStrategy}(this, name)
}
\arguments{
\item{this}{modelStrategy}

\item{name}{character, name of indicator}
}
\description{
Removes indicator by name
}
