% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{removeRule.modelStrategy}
\alias{removeRule.modelStrategy}
\title{Removes rule by name}
\usage{
\method{removeRule}{modelStrategy}(this, name)
}
\arguments{
\item{this}{modelStrategy}

\item{name}{character, name of rule}
}
\description{
Removes rule by name
}
