% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{removeStat}
\alias{removeStat}
\alias{removeStat.modelStrategy}
\title{Deletes variable from stats}
\usage{
removeStat(this, name)

\method{removeStat}{modelStrategy}(this, name)
}
\arguments{
\item{this}{modelStrategy}

\item{name}{character}
}
\description{
Deletes variable from stats
}
