% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOsFun.R
\name{sameMoneyOs}
\alias{sameMoneyOs}
\title{Function for volume in rule}
\usage{
sameMoneyOs(beta, amount, price, ...)
}
\arguments{
\item{beta}{numeric, number of assets}

\item{amount}{numeric, amount of money}

\item{price}{numeric, vector of prices}

\item{...}{args}
}
\value{
numeric
}
\description{
How many sum(price * abs(beta)) in amount
}
