% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{saveData.modelStrategy}
\alias{saveData.modelStrategy}
\title{Set names of data, that should be saved after finishing of perform function}
\usage{
\method{saveData}{modelStrategy}(this, s)
}
\arguments{
\item{this}{modelStrategy}

\item{s}{character, array of names}
}
\description{
Set names of data, that should be saved after finishing of perform function
}
