% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{scp_download}
\alias{scp_download}
\title{Download files from the server}
\usage{
scp_download(session, file, to = ".", ...)
}
\arguments{
\item{session}{ssh_session}

\item{file}{character, name of a file in the working directory of user}

\item{to}{character, path to folder where to upload file}

\item{...}{params}
}
\description{
Download files from the server
}
