% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{scp_upload}
\alias{scp_upload}
\title{Upload files to server}
\usage{
scp_upload(session, file, to = ".", ...)
}
\arguments{
\item{session}{ssh_session}

\item{file}{character, path to file}

\item{to}{character, name of folder to upload}

\item{...}{params}
}
\description{
Upload files to server
}
