% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setBetasInt.modelStrategy}
\alias{setBetasInt.modelStrategy}
\title{Sets if you have a need to convert betas to int.}
\usage{
\method{setBetasInt}{modelStrategy}(model, x)
}
\arguments{
\item{model}{modelStrategy}

\item{x}{logical}
}
\description{
Sets if you have a need to convert betas to int.
}
