% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setBetasInt}
\alias{setBetasInt}
\alias{setBetasInt.modelStrategy}
\title{Set TRUE if you have a need to convert betas to integers.}
\usage{
setBetasInt(this, x)

\method{setBetasInt}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{logical}
}
\description{
Set TRUE if you have a need to convert betas to integers.
}
