% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{setBetaTable.modelStrategy}
\alias{setBetaTable.modelStrategy}
\title{Set your coefficients for any moment}
\usage{
\method{setBetaTable}{modelStrategy}(this, table, force_fun = FALSE)
}
\arguments{
\item{this}{modelStrategy}

\item{table}{xts, columns should go in same order, that in user data}

\item{force_fun}{logical, if it is TRUE, then beta_fun will be specified 
if it is FALSE, then it will be specified only if it is NULL}
}
\description{
Set your coefficients for any moment
}
\examples{
tmp <- xts(data.frame(GAZP = c(10, 20), LKOH = c(3, 4)), c(as.Date('2010-01-01'), as.Date('2013-01-01')))
setBetaTable(this, tmp)
}
