% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission_slippage.R
\name{setCommission.modelStrategy}
\alias{setCommission.modelStrategy}
\title{Sets commission to model}
\usage{
\method{setCommission}{modelStrategy}(this, q)
}
\arguments{
\item{this}{modelStrategy}

\item{q}{quote, it should depend from pos_change argument, but it can include any variable from model}
}
\description{
Sets commission to model
}
