% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericModelStrategy.R, R/modelStrategy.R
\name{setExpandingLookback}
\alias{setExpandingLookback}
\alias{setExpandingLookback.modelStrategy}
\title{set expandingLookback}
\usage{
setExpandingLookback(this, x)

\method{setExpandingLookback}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{logical}
}
\description{
If it is TRUE, then size of data in beta_fun will be equal to current index, else to lookback
}
