% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setIgnorePosition.modelStrategy}
\alias{setIgnorePosition.modelStrategy}
\title{Gets variable ignorePosition.}
\usage{
\method{setIgnorePosition}{modelStrategy}(this, bool)
}
\arguments{
\item{this}{modelStrategy}

\item{bool}{bool, if it is true, then when model  simulated, when there was time to change coefs and we had position,
then position would be closed}
}
\description{
Gets variable ignorePosition.
}
