% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setLookback.modelStrategy}
\alias{setLookback.modelStrategy}
\title{Sets window for calculating coefs}
\usage{
\method{setLookback}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window for calculating}
}
\description{
Sets window for calculating coefs
}
