% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setLookForward}
\alias{setLookForward}
\title{Set window how far the same coefs will be used}
\usage{
setLookForward(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window}
}
\description{
Set window how far the same coefs will be used
}
