% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setMaxLookback}
\alias{setMaxLookback}
\alias{setMaxLookback.modelStrategy}
\title{Set window for calculating indicators}
\usage{
setMaxLookback(this, x)

\method{setMaxLookback}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window}
}
\description{
Set window for calculating indicators
}
