% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R
\name{setModelD.modelPortfolio}
\alias{setModelD.modelPortfolio}
\title{Set modelData}
\usage{
setModelD.modelPortfolio(this, data)
}
\arguments{
\item{this}{modelPortfolio}
}
\description{
Set modelData
}
