% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setModelD.modelStrategy}
\alias{setModelD.modelStrategy}
\title{Sets modelData object to modelStrategy object}
\usage{
\method{setModelD}{modelStrategy}(this, x, clearBacktests = TRUE)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{modelData}
}
\description{
Sets modelData object to modelStrategy object
}
