% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setMoney}
\alias{setMoney}
\alias{setMoney.modelStrategy}
\title{Set amount of money, that we have at the beginning}
\usage{
setMoney(this, x)

\method{setMoney}{modelStrategy}(this, x)
}
\arguments{
\item{this}{model}

\item{x}{numeric, amount of money}
}
\description{
Set amount of money, that we have at the beginning
}
