% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{setParams.modelStrategy}
\alias{setParams.modelStrategy}
\title{Set list of parameters}
\usage{
\method{setParams}{modelStrategy}(this, type, args)
}
\arguments{
\item{this}{modelStrategy}

\item{type}{character, it can be equal to rules, indicators, pp or pm}

\item{args}{list, define your params here}
}
\description{
Here you can define parameters for rules, indicators, program part and managers of position
}
\examples{
setParams(this,
       type = 'rules',
       args = list(n = 5, a = 2)
)
}
