% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setPmAfterOpen}
\alias{setPmAfterOpen}
\alias{setPmAfterOpen.modelStrategy}
\title{Gets pmAfterOpen.}
\usage{
setPmAfterOpen(this, x)

\method{setPmAfterOpen}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{logical}
}
\description{
If it is true, then position manager will be turned on just after opening position
}
