% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymodel.R
\name{setPyModelParams}
\alias{setPyModelParams}
\alias{setPyModelParams.modelStrategy}
\title{Set params of python model}
\usage{
setPyModelParams(this, ...)

\method{setPyModelParams}{modelStrategy}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}
}
\description{
Set params of python model
}
