% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setToleranceBeta.modelStrategy}
\alias{setToleranceBeta.modelStrategy}
\title{Sets tolerance of betas computation}
\usage{
\method{setToleranceBeta}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, tolerance}
}
\description{
Sets tolerance of betas computation
}
