% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setTradingData.modelStrategy}
\alias{setTradingData.modelStrategy}
\title{tradingData indicates which data strategy should use for spread calculating}
\usage{
\method{setTradingData}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
tradingData indicates which data strategy should use for spread calculating
}
