% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{setUserData}
\alias{setUserData}
\title{Select data to backtest.}
\usage{
setUserData(this, l)
}
\arguments{
\item{this}{modelStrategy}

\item{l}{list or xts, list can contain xts in each cell or 
should contain following fields:
1. dataset -- name of dataset, now only "Russia" is available
2. period -- periodicity of data. It is one of c('hour', 'day').
3. assets -- character vector of names of assets from dataset
4. time -- numeric, if period == 'day', then time should be specified. Time here is hour.
For example, if time == 19, then data will be at 19:00 every business day. This field works only in
Russia dataset. Xts -- table of timeserieses of instruments.
if l is xts, then each column should be close price for instruments.}
}
\description{
Select data to backtest.
}
\examples{
\dontrun{
setUserData(this, list(
   dataset = 'Russia',
   period = 'day',
   assets = c('GAZP', 'LKOH'),
   time = 19
))

x <- xts(cumsum(rnorm(1000)) + 300, Sys.Date()  + 1:1000)
setUserData(this, x)

x <- xts(cumsum(rnorm(1000)) + 300, Sys.Date()  + 1:1000)
y <- xts(cumsum(rnorm(1000)) + 300, Sys.Date()  + 1:1000)
setUserData(this, cbind(x, y))

x <- quantmod::getSymbols("AAPL", from = Sys.Date() - 365)
x <- quantmod::getSymbols("MSFT", from = Sys.Date() - 365)
setUserData(this, cbind(x, y))
}
}
