% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setWaitAfterClose.modelStrategy}
\alias{setWaitAfterClose.modelStrategy}
\title{Sets waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time}
\usage{
\method{setWaitAfterClose}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Sets waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time
}
