% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setWaitAfterClose}
\alias{setWaitAfterClose}
\alias{setWaitAfterClose.modelStrategy}
\title{Set waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time}
\usage{
setWaitAfterClose(this, x)

\method{setWaitAfterClose}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{logical}
}
\description{
Set waitAfterClose. If it is true, then new open will be after last close, else they may be in the same time
}
