% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Shiny_plot}
\alias{Shiny_plot}
\title{Plot interactive distribution params}
\usage{
Shiny_plot(this, ...)
}
\arguments{
\item{this}{it is Strategy}

\item{...}{params for shinyApp or paramset.index}
}
\description{
Plot interactive distribution params
}
