% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{ssh_connect}
\alias{ssh_connect}
\title{Connect to server}
\usage{
ssh_connect(host, keyfile = NULL, passwd = askpass::askpass,
  verbose = FALSE)
}
\arguments{
\item{host}{character, name of host}

\item{keyfile}{character, path to key}

\item{passwd}{askpass::askpass}

\item{verbose}{logical}
}
\description{
Connect to server
}
