% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{submit.modelStrategy}
\alias{submit.modelStrategy}
\title{Submit strategy to contest}
\usage{
\method{submit}{modelStrategy}(this, contest, session, verbose = FALSE)
}
\arguments{
\item{this}{modelStrategy}

\item{contest}{character, name of contest}

\item{session}{ssh_session}

\item{verbose}{logical}
}
\description{
Submit strategy to contest
}
