% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programParts.R
\name{updateStat.modelStrategy}
\alias{updateStat.modelStrategy}
\title{Updates all variables in stats from environment}
\usage{
\method{updateStat}{modelStrategy}(this, e)
}
\arguments{
\item{this}{modelStrategy}

\item{e}{environment}
}
\description{
Updates all variables in stats from environment
}
