% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_examples.R
\name{upload_example}
\alias{upload_example}
\title{Upload example to server}
\usage{
upload_example(session, file_path, file_name = NULL)
}
\arguments{
\item{session}{ssh_session}

\item{file_path}{character, path to file}

\item{file_name}{character, how to name your file on server,
if NULL, then name will be the same as in your machine}
}
\description{
Upload example to server
}
