% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymodel.R
\name{uploadPyModel.environment}
\alias{uploadPyModel.environment}
\title{Upload python model to server}
\usage{
\method{uploadPyModel}{environment}(this, session)
}
\arguments{
\item{this}{modelStrategy}

\item{session}{ssh_session}
}
\description{
Upload python model to server
}
