% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymodel.R
\name{uploadPyModel}
\alias{uploadPyModel}
\alias{uploadPyModel.modelStrategy}
\alias{uploadPyModel.environment}
\alias{uploadPyModel.modelPortfolio}
\title{Upload python model to server}
\usage{
uploadPyModel(this, session)

\method{uploadPyModel}{modelStrategy}(this, session)

\method{uploadPyModel}{environment}(this, session)

\method{uploadPyModel}{modelPortfolio}(this, session)
}
\arguments{
\item{this}{modelStrategy}

\item{session}{ssh_session}
}
\description{
Upload python model to server
}
