% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R,
%   R/distribution_contraints.R
\name{addDistribution.modelStrategy}
\alias{addDistribution.modelStrategy}
\alias{addDistribution.modelStrategy}
\title{Add distribution}
\usage{
\method{addDistribution}{modelStrategy}(this, paramset.label, component.type,
  component.label, variable, label)

\method{addDistribution}{modelStrategy}(this, paramset.label, component.type,
  component.label, variable, label)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, label of paramset}

\item{component.type}{character, one of rule, indicator, params, lookback, lookforward, beta_fun}

\item{component.label}{character, name of component, argument 'as' is resposible for that. 
If component.type is equal to one of lookback, lookforward, beta_fun, then this argument should be missed}

\item{variable}{list with only one element, for example list(n = 1:10)}

\item{label}{character, name for this distribution}

\item{this}{modelStrategy}

\item{paramset.label}{character, label of paramset}

\item{component.type}{character, one of rule, indicator, params, lookback, lookforward, beta_fun}

\item{component.label}{character, name of component, argument 'as' is resposible for that. 
If component.type is equal to one of lookback, lookforward, beta_fun, then this argument should be missed}

\item{variable}{list with only one element, for example list(n = 1:10)}

\item{label}{character, name for this distribution}
}
\description{
Add distribution

Add distribution
}
\examples{
addIndicator(this, args = list(name = BBands, HLC = quote(spread), n = 20, sd = 1), as = 'bb',
        lookback = 100)
addDistribution(this,
    paramset.label = paramset,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread,1) > Lag(bb[ ,'up'],1)) &
                  (spread < bb[, 'up']) &
                  (spread > bb[ ,'mavg']) &
                  (abs(spread - bb[, 'mavg'])  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short',
     osFun = stratbuilder2pub:::sameMoneyOs,
     osFun_args = alist(amount = 5000000))
addDistribution(this,
    paramset.label = paramset,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)
addIndicator(this, args = list(name = BBands, HLC = quote(spread), n = 20, sd = 1), as = 'bb',
        lookback = 100)
addDistribution(this,
    paramset.label = paramset,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread,1) > Lag(bb[ ,'up'],1)) &
                  (spread < bb[, 'up']) &
                  (spread > bb[ ,'mavg']) &
                  (abs(spread - bb[, 'mavg'])  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short',
     osFun = stratbuilder2pub:::sameMoneyOs,
     osFun_args = alist(amount = 5000000))
addDistribution(this,
    paramset.label = paramset,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)
}
