% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R,
%   R/distribution_contraints.R
\name{addDistributionConstraint.list}
\alias{addDistributionConstraint.list}
\title{Add distributions' constraint to list of models}
\usage{
\method{addDistributionConstraint}{list}(l, ...)

\method{addDistributionConstraint}{list}(l, ...)
}
\arguments{
\item{l}{list, list of modelStrategy objects}

\item{...}{params for addDistribution}

\item{l}{list, list of modelStrategy objects}

\item{...}{params for addDistribution}
}
\description{
This method add the same distributions' constraint to each model in list

This method add the same distributions' constraint to each model in list
}
