% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R
\name{addDistributionConstraint.modelStrategy}
\alias{addDistributionConstraint.modelStrategy}
\title{Adds contraints to distributions}
\usage{
\method{addDistributionConstraint}{modelStrategy}(this, paramset.label, expr,
  label)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, paramset label}

\item{expr}{expression, that contains names from labels of distributions}

\item{label}{character, name of the constraint}
}
\description{
Adds contraints to distributions
}
\examples{
addIndicator(this, args = list(name = BBands, HLC = quote(spread), n = 20, sd = 1), as = 'bb',
        lookback = 100)
addDistribution(this,
    paramset.label = paramset,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread,1) > Lag(bb[, 'up'],1)) &
                  (spread < bb[, 'up']) &
                  (spread > bb[, 'mavg']) &
                  (abs(spread - bb[ ,'mavg'])/spread  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short',
     osFun = stratbuilder2pub:::sameMoneyOs,
     osFun_args = alist(amount = 5000000))
addDistribution(this,
    paramset.label = paramset,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)

addDistributionConstraint(this,
    paramset.label = paramset,
    expr = bb.sd > my_distr * 100
)
}
