% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{applyParamsetServer.modelStrategy}
\alias{applyParamsetServer.modelStrategy}
\title{Do backtest of combinations from paramset}
\usage{
\method{applyParamsetServer}{modelStrategy}(this, session, paramset.label,
  nsamples = 100, verbose = FALSE, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{session}{ssh_session}

\item{paramset.label}{charater, name of paramset}

\item{nsamples}{numeric, how many backtests to do}

\item{start_date}{Date / charater in format 'yyyy-mm-dd', start date of simulation}

\item{end_date}{Date / charater in format 'yyyy-mm-dd', end date of simulation}
}
\value{
data.frame, reports of simulations
}
\description{
Do backtest of combinations from paramset
}
