% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{clearTradeTime.modelStrategy}
\alias{clearTradeTime.modelStrategy}
\title{Remove trading time by type}
\usage{
\method{clearTradeTime}{modelStrategy}(this, type = "all")
}
\arguments{
\item{this}{modelStrategy}

\item{type}{character, one of c('all', 'enter', 'open', 'exit', 'close')}
}
\description{
Remove trading time by type
}
