% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_constraints.R
\name{deleteParamset.modelStrategy}
\alias{deleteParamset.modelStrategy}
\title{Removes paramset from strategy}
\usage{
\method{deleteParamset}{modelStrategy}(this, paramset.label)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{name of paramset}
}
\description{
Removes paramset from strategy
}
