% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getBacktestResults}
\alias{getBacktestResults}
\alias{getBacktestResults.modelStrategy}
\alias{getBacktestResults.list}
\alias{getBacktestResults.modelPortfolio}
\title{Results of backtest}
\usage{
getBacktestResults(this, paramset.label)

\method{getBacktestResults}{modelStrategy}(this, paramset.label)

\method{getBacktestResults}{list}(this, ...)

\method{getBacktestResults}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{paramset.label}{character, name of paramset}

\item{...}{arguments to getBacktestResults.modelStrategy}
}
\value{
data.frame

list of data.frames

list of data.frames
}
\description{
Returns model statistics, see details
}
\details{
trades - Number of trade

trades.year - Average annual number of trades

long.trades - Number of long trade

short.trades - Number of short trade

long.acc - Long position accuracy

short.acc - Short position accuracy

total.acc - Position accuracy

max.loose - Max loose streak

max.win - Max win streak

return.ann - Average annual return

return.avg - Average annual return on capital employed

return.pos.drawdown  - PnL/(max(money in pos) + max.drawdow), where max.drawdow -> https://en.wikipedia.org/wiki/Drawdown_(economics)

drawdown.money - max.drawdown/money_in_strategy

median - median(profit_in_trade/money_in_strategy)

in.pos - time_in_position/ all_time

in.pos.positive  - (time_in_position where unrealized PnL > 0)/ time_in_position

days.in.pos.max - max number of days in position

days.in.pos.mean - mean number of days in position

days.out.pos.max - max number of days out of position

days.out.pos.mean - mean number of days out of position

sharpe.ann - sharpe coefficient https://en.wikipedia.org/wiki/Sharpe_ratio

sortino.ann - sortino coefficiet https://en.wikipedia.org/wiki/Sortino_ratio

maxMAE.money - Max absolute error in trade

profit.drawdown.year - PnL/max.drawdow
}
