% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCapital}
\alias{getCapital}
\alias{getCapital.modelStrategy}
\alias{getCapital.modelPortfolio}
\title{Return money in position}
\usage{
getCapital(this, ...)

\method{getCapital}{modelStrategy}(this, start_date = NULL,
  end_date = NULL, ...)

\method{getCapital}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{start_date}{date type, example: start_date='2008-01-01'}

\item{end_date}{date type, example: end_date='2018-01-01'}
}
\description{
Return money in position
}
