% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDateByIndex.modelStrategy}
\alias{getDateByIndex.modelStrategy}
\alias{getDateByIndex.modelPortfolio}
\title{Gets dates by indexes from modelData}
\usage{
\method{getDateByIndex}{modelStrategy}(this, indexes = NULL)

\method{getDateByIndex}{modelPortfolio}(this, indexes = NULL)
}
\arguments{
\item{this}{modelStrategy}

\item{indexes}{numeric vector or NULL}

\item{this}{modelPortfolio}

\item{indexes}{numeric vector or NULL}
}
\value{
vector of dates

vector of dates
}
\description{
Gets dates by indexes from modelData

Gets dates by indexes from modelData
}
