% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getIgnorePosition}
\alias{getIgnorePosition}
\alias{getIgnorePosition.modelStrategy}
\title{Get variable ignorePosition.}
\usage{
\method{getIgnorePosition}{modelStrategy}(this)

\method{getIgnorePosition}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
If it is true, then when model  simulated, when there was time to change coefs and we had position,
 then position would be closed
}
