% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{getIndicators}
\alias{getIndicators}
\alias{getIndicators.modelStrategy}
\title{Get inforamtion about indicators in the model}
\usage{
getIndicators(this)

\method{getIndicators}{modelStrategy}(this)
}
\arguments{
\item{this}{model}
}
\value{
list of inforamation
}
\description{
Get inforamtion about indicators in the model
}
