% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getLookback}
\alias{getLookback}
\alias{getLookback.modelStrategy}
\title{Get window for calculating coefs}
\usage{
getLookback(this)

\method{getLookback}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Get window for calculating coefs
}
