% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getLookForward.modelStrategy}
\alias{getLookForward.modelStrategy}
\title{Gets window how far the same coefs will be used}
\usage{
\method{getLookForward}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets window how far the same coefs will be used
}
