% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getMaxLookback.modelStrategy}
\alias{getMaxLookback.modelStrategy}
\title{Gets window for calculating indicators}
\usage{
\method{getMaxLookback}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets window for calculating indicators
}
