% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getModelD.modelStrategy}
\alias{getModelD.modelStrategy}
\title{Gets modelData object from modelStrategy object}
\usage{
\method{getModelD}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Gets modelData object from modelStrategy object
}
