% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getMoney.modelStrategy}
\alias{getMoney.modelStrategy}
\title{Gets amount of money that we have at the beginning in specific backtest}
\usage{
\method{getMoney}{modelStrategy}(this, from = NULL)
}
\arguments{
\item{this}{modelStrategy}

\item{from}{character, name of folder in backtests}
}
\value{
numeric, amount of money
}
\description{
Gets amount of money that we have at the beginning in specific backtest
}
