% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionManagers.R
\name{getPM}
\alias{getPM}
\alias{getPM.modelStrategy}
\title{Return all position managers blocks}
\usage{
getPM(this)

\method{getPM}{modelStrategy}(this)
}
\arguments{
\item{this}{model}
}
\value{
list
}
\description{
Return all position managers blocks
}
