% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPnL}
\alias{getPnL}
\alias{getPnL.modelStrategy}
\alias{getPnL.modelPortfolio}
\alias{getPnL.list}
\title{Get pnl as xts series}
\usage{
getPnL(this, ...)

\method{getPnL}{modelStrategy}(this, start_date = NULL,
  end_date = NULL, comOn = TRUE, leg = "all", ...)

\method{getPnL}{modelPortfolio}(this, ...)

\method{getPnL}{list}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{start_date}{date type, example: start_date='2008-01-01'}

\item{end_date}{date type, example: end_date='2018-01-01'}

\item{comOn}{bool, if true then commission will be included in the 'trades' graph}

\item{leg}{numeric/character, number of leg, if it is equal to 'all' or 'sum', then all pnl among all legs}
}
\value{
xts zoo

xts zoo
}
\description{
This method should be called only after backtest is done
}
