% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportCalendar}
\alias{getReportCalendar}
\alias{getReportCalendar.modelStrategy}
\alias{getReportCalendar.list}
\alias{getReportCalendar.modelPortfolio}
\title{Report of strategy year by year}
\usage{
getReportCalendar(this)

\method{getReportCalendar}{modelStrategy}(this)

\method{getReportCalendar}{list}(this)

\method{getReportCalendar}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list of data.frames
}
\description{
Returns model statistics, see details
}
\details{
sharpe.ann - sharpe coefficient https://en.wikipedia.org/wiki/Sharpe_ratio

return.pos.drawdown  - PnL/(max(money in pos) + max.drawdow), where max.drawdow -> https://en.wikipedia.org/wiki/Drawdown_(economics)

return.avg - Average annual return on capital employed

return.ann - Average annual return

return.drawdown.year - PnL/max.drawdow
}
