% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportTrades}
\alias{getReportTrades}
\alias{getReportTrades.modelStrategy}
\alias{getReportTrades.list}
\alias{getReportTrades.modelPortfolio}
\title{Report of trades}
\usage{
getReportTrades(this)

\method{getReportTrades}{modelStrategy}(this)

\method{getReportTrades}{list}(this)

\method{getReportTrades}{modelPortfolio}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Returns model statistics, see details
}
\details{
ind.start - index start trade

index.send - index end trede

date.start - date start trade

date.end - date end trade

side - position 1 - long, -1 - short

beta - beta spread

price.start - price in the date.start

price.end - price in the date.end

pnl.asset - pnl by asset

com.asset - commission by asset

com.sum - commissions on all assets

pnl.sum - PnL on all assets

MAE.with.com - max loss in position

MFE.with.com - max win in position

return.from.money - profit/(money_in_strategy)
}
