% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{getRules.modelStrategy}
\alias{getRules.modelStrategy}
\title{Gets inforamtion about rules in the model}
\usage{
\method{getRules}{modelStrategy}(this, pathwise = FALSE,
  recalc = FALSE)
}
\arguments{
\item{this}{modelStrategy}
}
\value{
list of inforamation
}
\description{
Gets inforamtion about rules in the model
}
