% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{getRules}
\alias{getRules}
\alias{getRules.modelStrategy}
\title{Get inforamtion about rules in the model}
\usage{
getRules(this, pathwise = FALSE, recalc = FALSE)

\method{getRules}{modelStrategy}(this, pathwise = FALSE, recalc = FALSE)
}
\arguments{
\item{this}{modelStrategy}

\item{pathwise}{logical, if TRUE only pathwise rules will be returned}

\item{recalc}{logical, if TRUE then rules will be recalculated}
}
\value{
list of inforamation
}
\description{
Get inforamtion about rules in the model
}
