% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{getSavedModels}
\alias{getSavedModels}
\alias{getSavedModels.modelStrategy}
\title{Return Saved Models}
\usage{
getSavedModels(this)

\method{getSavedModels}{modelStrategy}(this)
}
\arguments{
\item{this}{modelStrategy}
}
\description{
Return Saved Models
}
