% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{getTradeTime}
\alias{getTradeTime}
\alias{getTradeTime.modelStrategy}
\title{Gets tarding time of strategy}
\usage{
getTradeTime(this, type = "all")

\method{getTradeTime}{modelStrategy}(this, type = "all")
}
\arguments{
\item{this}{model}

\item{type}{character, one of  c('all', 'enter', 'open', 'exit', 'close')}
}
\description{
Gets tarding time of strategy
}
