% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_communication.R
\name{performServer.list}
\alias{performServer.list}
\title{Simulate list of strategies}
\usage{
\method{performServer}{list}(l, session, verbose = FALSE, ...)
}
\arguments{
\item{l}{list, list of strategies}

\item{session}{ssh_session}

\item{...}{additional params}

\item{reports}{character vector, it can include c('strategy', 'calendar', 'trades', 'plot')}
}
\value{
list
}
\description{
Simulate list of strategies
}
