% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCapital}
\alias{plotCapital}
\alias{plotCapital.modelStrategy}
\alias{plotCapital.modelPortfolio}
\title{Plot Capital of strategy}
\usage{
plotCapital(this, ...)

\method{plotCapital}{modelStrategy}(this, interactive_plot = TRUE,
  start_date = NULL, end_date = NULL, ...)

\method{plotCapital}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{interactive_plot, }{logical, use FALSE if you want to disable the interactive graph}

\item{start_date}{date type, example: start_date='2008-01-01'}

\item{end_date}{date type, example: end_date='2018-01-01'}
}
\description{
Plot Capital of strategy
}
