% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/plots.R
\name{plotDrawdowns.modelPortfolio}
\alias{plotDrawdowns.modelPortfolio}
\alias{plotDrawdowns}
\alias{plotDrawdowns}
\alias{plotDrawdowns.modelStrategy}
\alias{plotDrawdowns.list}
\title{Plot drawdowns}
\usage{
\method{plotDrawdowns}{modelPortfolio}(this, ...)

plotDrawdowns(this, ...)

plotDrawdowns(this, ...)

\method{plotDrawdowns}{modelStrategy}(this, return_type = "plot",
  graph_type = "ggplot2", interactive_plot = FALSE, ...)

\method{plotDrawdowns}{list}(this, legend = TRUE, interactive_plot = FALSE,
  ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{return_type}{character, plot or data}

\item{graph_type}{character, ggplot2 or xts}

\item{interactive_plot}{logical, if graph_type == 'ggplot2' and this option is TRUE, then plot will be interactive}

\item{legend}{logical, if true then legend will be printed on the plot}

\item{this}{modelStrategy}

\item{...}{params}
}
\value{
ggplot/xts

ggplot/xts
}
\description{
Plot drawdowns

Plot drawdowns
}
