% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDrawdowns}
\alias{plotDrawdowns}
\alias{plotDrawdowns.modelStrategy}
\alias{plotDrawdowns.list}
\alias{plotDrawdowns.modelPortfolio}
\title{Plot drawdowns}
\usage{
plotDrawdowns(this, ...)

\method{plotDrawdowns}{modelStrategy}(this, return_type = "plot",
  graph_type = "ggplot2", ...)

\method{plotDrawdowns}{list}(this, legend = TRUE, ...)

\method{plotDrawdowns}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{return_type}{character, plot or data}

\item{graph_type}{character, ggplot2 or xts}

\item{legend}{logical, if true then legend will be printed on the plot}
}
\value{
ggplot/xts
}
\description{
Plot drawdowns
}
