% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPnL}
\alias{plotPnL}
\alias{plotPnL.modelStrategy}
\alias{plotPnL.list}
\alias{plotPnL.modelPortfolio}
\title{Plot change of pnl through backtest period}
\usage{
plotPnL(this, ...)

\method{plotPnL}{modelStrategy}(this, type = "money", comOn = TRUE,
  leg = "all", graph_type = "ggplot2", each_year = FALSE,
  adjust = FALSE, return_type = "plot", cutoff = FALSE, ...)

\method{plotPnL}{list}(this, legend = TRUE, ...)

\method{plotPnL}{modelPortfolio}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{...}{params}

\item{type}{character, one of c('money','trades','percents')}

\item{comOn}{bool, if true then commission will be included in the 'trades' graph}

\item{leg}{numeric/character, number of leg, if it is equal to 'all' or 'sum', then all pnl among all legs
will be summed, if it is equal to 'sep', then pnl among legs will be plotted}

\item{graph_type}{character, ggplot2 or xts}

\item{each_year}{logical, if TRUE, then each graph will start with 0 each year}

\item{adjust}{logical, if TRUE, then values will be divided by getMoney(this)}

\item{return_type}{character, plot or data}

\item{cutoff}{logical, if TRUE then on plot will be horizonal line indicating when model was created}

\item{legend}{logical, if true then legend will be printed on the plot}
}
\description{
Plot change of pnl through backtest period
}
