% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotReturns}
\alias{plotReturns}
\alias{plotReturns.modelStrategy}
\title{Plot returns vs MAE/MFE}
\usage{
plotReturns(this, type = "MAE")

\method{plotReturns}{modelStrategy}(this, type = "MAE")
}
\arguments{
\item{this}{modelStrategy}

\item{type}{character, MAE or MFE}
}
\value{
ggplot
}
\description{
Plot returns vs MAE/MFE
}
