% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotShiny}
\alias{plotShiny}
\alias{plotShiny.modelStrategy}
\title{Plot interactive distribution params}
\usage{
plotShiny(this, ...)

\method{plotShiny}{modelStrategy}(this, session, paramset = 1,
  delite_save = 0, start_date = "1900-01-01",
  end_date = "2999-01-01", ...)
}
\arguments{
\item{this}{it is Strategy}

\item{...}{params for shinyApp}

\item{session}{object of class ssh_session}

\item{paramset}{name of paramset}

\item{delite_save}{please use 1, if you want delite save strategy}

\item{start_date}{initial start date}

\item{end_date}{initial stop date}
}
\description{
Plot interactive distribution params
}
