% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStrategy}
\alias{plotStrategy}
\title{Plot opening and closing position}
\usage{
plotStrategy(this, multi_plot = F)
}
\arguments{
\item{this}{modelStrategy}

\item{multi_plot}{logical, if TRUE plot spread and legs}
}
\description{
Plot opening and closing position
}
