% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOsFun.R
\name{proportionOs}
\alias{proportionOs}
\title{Weight's functions}
\usage{
proportionOs(beta, money, proportion, price, ...)
}
\arguments{
\item{beta}{coefficients of spread}

\item{money}{all money that strategy has}

\item{proportion}{how much money we can use in rule}

\item{price}{current price of instruments}

\item{...}{additional args}
}
\value{
numeric multiplier( how much minimal position to buy)
}
\description{
Weight's functions
}
