% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOsFun.R
\name{proportionOs}
\alias{proportionOs}
\title{beta -- coefficients in spread
money -- all money that strategy has
proportion -- how much money we can use in rule
price -- current price of instruments( how much it costs to buy or sell)
it returns multiplier( how much minimal position )}
\usage{
proportionOs(beta, money, proportion, price, ...)
}
\description{
beta -- coefficients in spread
money -- all money that strategy has
proportion -- how much money we can use in rule
price -- current price of instruments( how much it costs to buy or sell)
it returns multiplier( how much minimal position )
}
