% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{removeIndicator}
\alias{removeIndicator}
\alias{removeIndicator.modelStrategy}
\title{Remove indicator by name}
\usage{
removeIndicator(this, name)

\method{removeIndicator}{modelStrategy}(this, name)
}
\arguments{
\item{this}{modelStrategy}

\item{name}{character, name of indicator}
}
\description{
Remove indicator by name
}
