% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setBeta.modelStrategy}
\alias{setBeta.modelStrategy}
\title{Sets rule for calculating spread.}
\usage{
\method{setBeta}{modelStrategy}(this, fun, args = NULL)
}
\arguments{
\item{this}{modelStrategy}

\item{fun}{function, it calculates coefs for spread}
}
\description{
Sets rule for calculating spread.
}
