% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setBeta}
\alias{setBeta}
\title{Set rule for calculating spread.}
\usage{
setBeta(this, fun, args = NULL)
}
\arguments{
\item{this}{modelStrategy}

\item{fun}{function, it calculates coefs for spread}

\item{args}{list, arguments in setBeta function}
}
\description{
Set rule for calculating spread.
}
