% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setBeta}
\alias{setBeta}
\alias{setBeta.modelStrategy}
\title{Set rule for calculating spread.}
\usage{
setBeta(this, fun, args = NULL, lookback, by_money)

\method{setBeta}{modelStrategy}(this, fun, args = NULL, lookback, by_money)
}
\arguments{
\item{this}{modelStrategy}

\item{fun}{function, it calculates coefs for spread}

\item{args}{list, arguments in setBeta function}

\item{lookback}{numeric, lookback for calculating coefficients, how many periods should be in data}

\item{by_money}{logical, if it is TRUE then coefficients will be interpreted as money position in instruments, else in pieces}

\item{this}{modelStrategy}

\item{fun}{function / numeric vector, it calculates coefs for spread}

\item{args}{list/NULL, default arguments for fun}
}
\description{
Set rule for calculating spread.

Set rule for calculating spread.
}
