% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setIgnorePosition}
\alias{setIgnorePosition}
\alias{setIgnorePosition.modelStrategy}
\title{Get variable ignorePosition.}
\usage{
setIgnorePosition(this, bool)

\method{setIgnorePosition}{modelStrategy}(this, bool)
}
\arguments{
\item{this}{modelStrategy}

\item{bool}{bool, if it is true, then when model  simulated, 
when there was time to change coefs and we had position,
 then position would be closed}
}
\description{
Get variable ignorePosition.
}
