% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setLookback}
\alias{setLookback}
\title{Set window for calculating coefs}
\usage{
setLookback(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window for calculating}
}
\description{
Set window for calculating coefs
}
