% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setLookForward.modelStrategy}
\alias{setLookForward.modelStrategy}
\title{Sets window how far the same coefs will be used}
\usage{
\method{setLookForward}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window}
}
\description{
Sets window how far the same coefs will be used
}
