% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setMaxLookback.modelStrategy}
\alias{setMaxLookback.modelStrategy}
\title{Sets window for calculating indicators}
\usage{
\method{setMaxLookback}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, window}
}
\description{
Sets window for calculating indicators
}
