% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/utils.R
\name{setModelD.modelPortfolio}
\alias{setModelD.modelPortfolio}
\alias{setModelD.modelPortfolio}
\title{Set modelData}
\usage{
\method{setModelD}{modelPortfolio}(this, data)

\method{setModelD}{modelPortfolio}(this, data)
}
\arguments{
\item{this}{modelPortfolio}

\item{this}{modelPortfolio}
}
\description{
Set modelData

Set modelData
}
