% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/utils.R
\name{setModelD.modelPortfolio}
\alias{setModelD.modelPortfolio}
\alias{setModelD.list}
\alias{setModelD}
\alias{setModelD.modelStrategy}
\title{Set modelData object to modelStrategy object}
\usage{
\method{setModelD}{modelPortfolio}(this, x, ...)

\method{setModelD}{list}(this, x, ...)

setModelD(this, x, clearBacktests = TRUE)

\method{setModelD}{modelStrategy}(this, x, clearBacktests = TRUE)

\method{setModelD}{modelPortfolio}(this, x, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{modelData}

\item{...}{params}

\item{clearBacktests}{logical, if TRUE then all backtests will be deleted}
}
\description{
Set modelData object to modelStrategy object
}
