% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStrategy.R
\name{setMoney.modelStrategy}
\alias{setMoney.modelStrategy}
\title{Sets amount of money, that we have at the beginning}
\usage{
\method{setMoney}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric, amount of money}
}
\description{
Sets amount of money, that we have at the beginning
}
