% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPortfolio.R, R/modelStrategy.R
\name{setMoney.modelPortfolio}
\alias{setMoney.modelPortfolio}
\alias{setMoney}
\alias{setMoney.modelStrategy}
\title{Set amount of money to modelPortfolio object.}
\usage{
\method{setMoney}{modelPortfolio}(this, x)

setMoney(this, x)

\method{setMoney}{modelStrategy}(this, x)
}
\arguments{
\item{this}{modelStrategy}

\item{x}{numeric type, new amount of money in strategy}

\item{this}{model}

\item{x}{numeric, amount of money}
}
\description{
It will spread that value across models inside it with old proportion.

Set amount of money, that we have at the beginning
}
