% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_examples.R
\name{update_package}
\alias{update_package}
\title{Update this package}
\usage{
update_package(user, keyfile, session = NULL)
}
\arguments{
\item{user}{character, ssh address}

\item{keyfile}{character, path to private key}

\item{session}{ssh_sessin}
}
\description{
Update this package
}
